python_screening_task/requirements.txt
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.2
urllib3==2.2.3